src/webapp/requirements.txt
fastapi
uvicorn
jinja2
azure-identity
azure-search-documents
openai
python-multipart
